refactor: move mode to model
