+++ src/model/Mode.ts
@@ -1,0 +1,3 @@
+export type Mode = "light" | "dark";
+
+export const defaultMode = "light";
