+++ src/repository/ThemeRepository.ts
@@ -1,9 +1,9 @@
-import { Mode } from "../view/contexts/ModeContext";
+import { Mode, defaultMode } from "../model/Mode";
 import { ChromeLocalStorage } from "./ChromeStorage";
 
 export const getMode = async (): Promise<Mode> => {
   const mode = await ChromeLocalStorage.getMode();
-  if (!mode) return "light";
+  if (!mode) return defaultMode;
 
   return mode;
 };
