+++ src/view/contexts/ModeContext.tsx
@@ -1,9 +1,8 @@
 import { ThemeProvider } from "@emotion/react";
 import { createTheme } from "@mui/material/styles";
 import { createContext, useEffect, useMemo, useState } from "react";
+import { Mode } from "../../model/Mode";
 import { getMode } from "../../repository/ThemeRepository";
-
-export type Mode = "light" | "dark";
 
 type ModeContextType = {
   mode: Mode;
